requirements.txt
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.0
Flask==3.1.1
Flask-SocketIO==5.5.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
requests==2.32.3
selenium==4.33.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
